requirements.txt
absl-py==0.11.0
astor==0.8.1
backcall==0.2.0
cached-property==1.5.2
colorama==0.4.4
cycler==0.10.0
decorator==4.4.2
future==0.18.2
gast==0.2.2
google-pasta==0.2.0
grpcio==1.34.0
h5py==3.1.0
imageio==2.9.0
importlib-metadata==3.1.1
ipython==7.19.0
ipython-genutils==0.2.0
jedi==0.17.2
Keras==2.1.2
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.2
kiwisolver==1.3.1
Markdown==3.3.3
matplotlib==3.3.3
networkx==2.5
numpy==1.18.5
opencv-python==4.4.0.46
opt-einsum==3.3.0
pandas==1.1.5
parso==0.7.1
pickleshare==0.7.5
Pillow==8.0.1
prompt-toolkit==3.0.8
protobuf==3.14.0
Pygments==2.7.3
pyparsing==2.4.7
python-dateutil==2.8.1
pytz==2020.4
PyWavelets==1.1.1
PyYAML==5.3.1
scikit-image==0.17.2
scipy==1.5.4
six==1.15.0
tensorboard==1.15.0
tensorflow-directml==1.15.3.dev200911
tensorflow-estimator==1.15.1
termcolor==1.1.0
tifffile==2020.12.4
torch==1.5.1+cpu
torchvision==0.6.1+cpu
traitlets==5.0.5
wcwidth==0.2.5
Werkzeug==1.0.1
wrapt==1.12.1
zipp==3.4.0
